fit-model/requirements.txt
pandas==1.1.3
numpy==1.18.1
scikit-learn==0.23.0
sklearn2pmml==0.63.1